Release 0.6.0

+++ gbm-sys/src/platforms/linux/arm/gen.rs
@@ -6,11 +6,11 @@ pub const GBM_BO_IMPORT_EGL_IMAGE: u32 = 21762;
 pub const GBM_BO_IMPORT_FD: u32 = 21763;
 pub const GBM_BO_IMPORT_FD_MODIFIER: u32 = 21764;
 pub const GBM_MAX_PLANES: u32 = 4;
-pub type size_t = libc::c_ulong;
+pub type size_t = libc::c_uint;
 pub type __int32_t = libc::c_int;
 pub type __uint32_t = libc::c_uint;
-pub type __int64_t = libc::c_long;
-pub type __uint64_t = libc::c_ulong;
+pub type __int64_t = libc::c_longlong;
+pub type __uint64_t = libc::c_ulonglong;
 #[doc = " \\file gbm.h"]
 #[doc = " \\brief Generic Buffer Manager"]
 #[repr(C)]
